requirements.txt
absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.2.7
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
branca==0.6.0
bs4==0.0.1
cachetools==5.3.1
certifi==2020.4.5.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.1
cryptography==40.0.2
cycler==0.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==3.0.6
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.41.0
fqdn==1.5.1
future==0.18.3
gast==0.4.0
geocoder==1.38.1
geographiclib==2.0
geolocation-python==0.2.2
geopy==2.3.0
gmplot==1.4.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==2.9
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.12.0
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
jwt==1.3.1
keras==2.12.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
ml-dtypes==0.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
panda==0.3.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdfminer==20191125
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.5.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.4
PyMuPDF==1.22.5
pyparsing==3.1.1
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-time==0.3.0
pytz==2020.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
ratelim==0.1.6
requests==2.23.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sockets==1.0.0
soupsieve==2.4.1
sqlparse==0.3.1
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2022.7
uri-template==1.2.0
urllib3==1.25.9
virtualenv==20.23.1
voila==0.4.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.14.1
